+++ client/src/components/app.tsx
@@ -1,7 +1,7 @@
 import { HomePage } from '@/pages/home-page';
 import { LoginPage } from '@/pages/login-page';
 import { ProtectedRoute } from '@/pages/protected-route';
-import { RegisterPage } from '@/pages/register';
+import { RegisterPage } from '@/pages/register-page';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import { BrowserRouter, Route, Routes } from "react-router";
@@ -12,7 +12,6 @@ import { ThemeProvider } from './theme-provider';
 const queryClient = new QueryClient()
 
 export function App() {
-
   return (
     <QueryClientProvider client={queryClient}>
       <ThemeProvider>
