+++ client/src/features/user-ranking/user-ranking-list.tsx
@@ -16,9 +16,5 @@ export function UserRankingList() {
 
   if (rankings?.length === 0) return <JoinLeagueCard />
 
-  return (
-    <div className='grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-4'>
-      {rankings?.map((ranking) => <RankingListItem key={ranking.id} ranking={ranking} />)}
-    </div>
-  )
+  return <>{rankings?.map((ranking) => <RankingListItem key={ranking.id} ranking={ranking} />)}</>
 }
