+++ client/src/pages/register-page.tsx
@@ -1,0 +1,9 @@
+import { RegisterCard } from '@/components/register-card'
+
+export function RegisterPage() {
+  return (
+    <div className='flex grow items-center justify-center px-4'>
+      <RegisterCard />
+    </div>
+  )
+}
